feat: add protection for new, edit e delete entry only access for usser logged

+++ learning_log/learning_logs/views.py
@@ -1,4 +1,5 @@
-from django.http import HttpResponseRedirect
+from django.contrib.auth.decorators import login_required
+from django.http import Http404, HttpResponseRedirect
 from django.shortcuts import render
 from django.urls import reverse
 
@@ -13,21 +14,28 @@ def index(request):
     return render(request, "learning_logs/index.html")
 
 
+@login_required
 def topics(request):
     """Mostra os assuntos"""
-    topics = Topic.objects.order_by("date_added")
+    topics = Topic.objects.filter(owner=request.user).order_by("date_added")
     context = {"topics": topics}
     return render(request, "learning_logs/topics.html", context)
 
 
+@login_required
 def topic(request, topic_id):
     """Mostra um único assunto com suas entradas"""
     topic = Topic.objects.get(id=topic_id)
-    entries = topic.entry_set.order_by("-date_added")
-    context = {"topic": topic, "entries": entries}
-    return render(request, "learning_logs/topic.html", context)
+    # Somente permite ver anotações do tópico do usuário logado
+    if topic.owner != request.user:
+        raise Http404
+    else:
+        entries = topic.entry_set.order_by("-date_added")
+        context = {"topic": topic, "entries": entries}
+        return render(request, "learning_logs/topic.html", context)
 
 
+@login_required
 def new_topic(request):
     """Adiciona um novo assunto."""
     if request.method != "POST":
@@ -37,6 +45,8 @@ def new_topic(request):
         # Dados de POST submetidos
         form = TopicForm(request.POST)
         if form.is_valid():
+            new_topic = form.save(commit=False)
+            new_topic.owner = request.user
             form.save()
             return HttpResponseRedirect(reverse("topics"))
 
@@ -44,53 +54,66 @@ def new_topic(request):
     return render(request, "learning_logs/new_topic.html", context)
 
 
+@login_required
 def new_entry(request, topic_id):
     """Adiciona uma nova anotação"""
     topic = Topic.objects.get(id=topic_id)
 
-    if request.method != "POST":
-        # Nenhum dado submetido cria um formulário em branco
-        form = Entryform()
+    if topic.owner != request.user:
+        raise Http404
     else:
-        # Dados de POST submetidos
-        form = Entryform(data=request.POST)
-        if form.is_valid():
-            new_entry = form.save(commit=False)
-            new_entry.topic = topic
-            new_entry.save()
-            return HttpResponseRedirect(reverse("topic", args=[topic_id]))
+        if request.method != "POST":
+            # Nenhum dado submetido cria um formulário em branco
+            form = Entryform()
+        else:
+            # Dados de POST submetidos
+            form = Entryform(data=request.POST)
+            if form.is_valid():
+                new_entry = form.save(commit=False)
+                new_entry.topic = topic
+                new_entry.save()
+                return HttpResponseRedirect(reverse("topic", args=[topic_id]))
 
-    context = {"topic": topic, "form": form}
-    return render(request, "learning_logs/new_entry.html", context)
+        context = {"topic": topic, "form": form}
+        return render(request, "learning_logs/new_entry.html", context)
 
 
+@login_required
 def edit_entry(request, entry_id):
     """Edita uma anotação"""
     entry = Entry.objects.get(id=entry_id)
     topic = entry.topic
 
-    if request.method != "POST":
-        # Requisição inicial para criar o formulário preenchido
-        form = Entryform(instance=entry)
+    if topic.owner != request.user:
+        raise Http404
 
     else:
-        # Dados de Post submetidos
-        form = Entryform(instance=entry, data=request.POST)
-        if form.is_valid():
-            form.save()
-            return HttpResponseRedirect(reverse("topic", args=[topic.id]))
+        if request.method != "POST":
+            # Requisição inicial para criar o formulário preenchido
+            form = Entryform(instance=entry)
 
-    context = {"entry": entry, "topic": topic, "form": form}
-    return render(request, "learning_logs/edit_entry.html", context)
+        else:
+            # Dados de Post submetidos
+            form = Entryform(instance=entry, data=request.POST)
+            if form.is_valid():
+                form.save()
+                return HttpResponseRedirect(reverse("topic", args=[topic.id]))
+
+        context = {"entry": entry, "topic": topic, "form": form}
+        return render(request, "learning_logs/edit_entry.html", context)
 
 
+@login_required
 def delete_entry(request, entry_id):
     entry = Entry.objects.get(id=entry_id)
     topic = entry.topic
 
-    if request.method == "POST":
-        entry.delete()
-        return HttpResponseRedirect(reverse("topic", args=[topic.id]))
+    if topic.owner != request.user:
+        raise Http404
+    else:
+        if request.method == "POST":
+            entry.delete()
+            return HttpResponseRedirect(reverse("topic", args=[topic.id]))
 
-    context = {"entry": entry, "topic": topic}
-    return render(request, "learning_logs/delete_entry.html", context)
+        context = {"entry": entry, "topic": topic}
+        return render(request, "learning_logs/delete_entry.html", context)
